+++ lib/plezi.rb
@@ -42,6 +42,7 @@ require 'plezi/helpers/mime_types.rb'
 require 'plezi/handlers/http_router.rb'
 require 'plezi/handlers/route.rb'
 require 'plezi/handlers/controller_magic.rb'
+require 'plezi/handlers/controller_core.rb'
 require 'plezi/handlers/stubs.rb'
