+++ lib/plezi/handlers/route.rb
@@ -256,157 +256,9 @@ module Plezi
 			new_class_name = "Plezi__#{controller.name.gsub /[\:\-\#\<\>\{\}\(\)\s]/, '_'}"
 			return Module.const_get new_class_name if Module.const_defined? new_class_name
 			# controller.include Plezi::ControllerMagic
-			controller.instance_exec(container) {|r| include Plezi::ControllerMagic; set_pl_route r;}
+			controller.instance_exec(container) {|r| include Plezi::ControllerMagic; }
 			ret = Class.new(controller) do
-
-				# def name
-				# 	new_class_name
-				# end
-
-				def initialize request, response
-					@request = request
-					@params = request.params
-					@flash = response.flash
-					@host_params = request.io[:params]
-					@response = response
-					@cookies = request.cookies
-					# @response["content-type"] ||= ::Plezi.default_content_type
-					@_accepts_broadcast = false
-					super()
-				end
-
-				# WebSockets.
-				#
-				# this method handles the protocol and handler transition between the HTTP connection
-				# (with a protocol instance of HTTPProtocol and a handler instance of HTTPRouter)
-				# and the WebSockets connection
-				# (with a protocol instance of WSProtocol and an instance of the Controller class set as a handler)
-				def pre_connect
-					# make sure this is a websocket controller
-					return false unless self.class.available_routing_methods.include?(:on_message)
-					# call the controller's original method, if exists, and check connection.
-					return false if (defined?(super) && !super)
-					# finish if the response was sent
-					return true if response.headers_sent?
-					# complete handshake
-					return true
-				end
-				# handles websocket opening.
-				def on_open ws
-					# set broadcasts and return true
-					@_accepts_broadcast = true
-					@response = ws
-					ws.autopong = 45 ###FIX THIS
-					# create the redis connection (in case this in the first instance of this class)
-					self.class.redis_connection
-					super() if defined?(super)
-				end
-				# def on_message ws
-				# 	super(ws.data) if defined?(super)
-				# end
-				# WebSockets.
-				#
-				# stops broadcasts from being called on closed sockets that havn't been collected by the garbage collector.
-				def on_close ws
-					@_accepts_broadcast = false
-					super() if defined? super
-				end
-
-				# Inner Routing
-				#
-				#
-				def _route_path_to_methods_and_set_the_response_
-					#run :before filter
-					return false if self.class.available_routing_methods.include?(:before) && self.before == false 
-					#check request is valid and call requested method
-					ret = requested_method
-					return false unless self.class.available_routing_methods.include?(ret)
-					ret = self.method(ret).call
-					return false unless ret
-					#run :after filter
-					return false if self.class.available_routing_methods.include?(:after) && self.after == false
-					# review returned type for adding String to response
-					return ret
-				end
-				# a callback that resets the class router whenever a method (a potential route) is added
-				def self.method_added(id)
-					reset_routing_cache
-				end
-				# a callback that resets the class router whenever a method (a potential route) is removed
-				def self.method_removed(id)
-					reset_routing_cache
-				end
-				# a callback that resets the class router whenever a method (a potential route) is undefined (using #undef_method).
-				def self.method_undefined(id)
-					reset_routing_cache
-				end
-
-				# # lists the available methods that will be exposed to HTTP requests
-				# def self.available_public_methods
-				# 	# set class global to improve performance while checking for supported methods
-				# 	@@___available_public_methods___ ||= available_routing_methods - [:before, :after, :save, :show, :update, :delete, :initialize, :on_message, :pre_connect, :on_connect, :on_disconnect]
-				# end
-
-				# # lists the available methods that will be exposed to the HTTP router
-				# def self.available_routing_methods
-				# 	# set class global to improve performance while checking for supported methods
-				# 	@@___available_routing_methods___ ||= (((public_instance_methods - Object.public_instance_methods) - Plezi::ControllerMagic::InstanceMethods.instance_methods).delete_if {|m| m.to_s[0] == '_'})
-				# end
-
-				# # resets this controller's router, to allow for dynamic changes
-				# def self.reset_routing_cache
-				# 	@@___available_routing_methods___ = @@___available_public_methods___ = nil
-				# 	available_routing_methods
-				# 	available_public_methods
-				# end
-
-				
-				# # reviews the Redis connection, sets it up if it's missing and returns the Redis connection.
-				# #
-				# # todo: review thread status? (incase an exception killed it)
-				# def self.redis_connection
-				# 	return false unless defined?(Redis) && ENV['PL_REDIS_URL']
-				# 	return @@redis if defined?(@@redis_sub_thread) && @@redis
-				# 	@@redis_uri ||= URI.parse(ENV['PL_REDIS_URL'])
-				# 	@@redis ||= Redis.new(host: @@redis_uri.host, port: @@redis_uri.port, password: @@redis_uri.password)
-				# 	@@redis_sub_thread = Thread.new do
-				# 		begin
-				# 			Redis.new(host: @@redis_uri.host, port: @@redis_uri.port, password: @@redis_uri.password).subscribe(redis_channel_name) do |on|
-				# 				on.message do |channel, msg|
-				# 					args = JSON.parse(msg)
-				# 					params = args.shift
-				# 					__inner_process_broadcast params['_pl_ignore_object'], params['_pl_method_broadcasted'].to_sym, args
-				# 				end
-				# 			end						
-				# 		rescue Exception => e
-				# 			Plezi.error e
-				# 			retry
-				# 		end
-				# 	end
-				# 	raise "Redis connction failed for: #{ENV['PL_REDIS_URL']}" unless @@redis
-				# 	@@redis
-				# end
-
-				# # returns a Redis channel name for this controller.
-				# def self.redis_channel_name
-				# 	self.name.to_s
-				# end
-
-				# # broadcasts messages (methods) for this process
-				# def self.__inner_process_broadcast ignore, method_name, args, &block
-				# 	ObjectSpace.each_object(self) { |controller| Plezi.callback controller, method_name, *args, &block if controller.accepts_broadcast? && (!ignore || controller.uuid != ignore) }
-				# end
-
-				# # broadcasts messages (methods) between all processes (using Redis).
-				# def self.__inner_redis_broadcast ignore, method_name, args, &block
-				# 	return false unless redis_connection
-				# 	raise "Radis broadcasts cannot accept blocks (no inter-process callbacks of memory sharing)!" if block
-				# 	# raise "Radis broadcasts accept only one paramater, which is an optional Hash (no inter-process memory sharing)" if args.length > 1 || (args[0] && !args[0].is_a?(Hash))
-				# 	args.unshift ({_pl_method_broadcasted: method_name, _pl_ignore_object: ignore})
-				# 	redis_connection.publish(redis_channel_name, args.to_json )
-				# 	true
-				# end
-
+				include Plezi::Base::ControllerCore
 			end
 			Object.const_set(new_class_name, ret)
 			Module.const_get(new_class_name).reset_routing_cache
